move EnvVariables one level up
