+++ ui/src/pages/namespace/Services/List/Row/index.tsx
@@ -10,7 +10,7 @@ import { TableCell, TableRow } from "~/design/Table";
 
 import Button from "~/design/Button";
 import { DialogTrigger } from "~/design/Dialog";
-import EnvsVariables from "./EnvVariables";
+import EnvsVariables from "../../components/EnvVariables";
 import { FC } from "react";
 import { ServiceSchemaType } from "~/api/services/schema/services";
 import { StatusBadge } from "../../components/StatusBadge";
