+++ ui/src/pages/namespace/Services/components/EnvVariables.tsx
